server/requirements.txt
connexion
coverage
Flask-SQLAlchemy
Flask-HTTPAuth
passlib
psycopg2
ipython
pytest
pytest-quickcheck
pytest-cov
